+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/AlbumTagsPanelBuilder.java
@@ -97,7 +97,6 @@ final class AlbumTagsPanelBuilder extends AbstractComponentValueBuilder<AlbumTag
 
 		AlbumTagsPanel() {
 			super(borderLayout());
-			System.out.println("new tag panel");
 			add(createCenterPanel(), BorderLayout.CENTER);
 			setupKeyEvents();
 		}